SQL Batch File + New RevisableDBOperator method
Sorry to be misleading on that other commit message. I don't actually have all methods implemented in RevisableDBOperator (still have retrieve with filter and revert operations to implement). There's also an SQL batch file that I plan to have a method that reads the file and executes on SQL server to reconstruct the schema if it doesn't exist in a system.

Signed-off-by: Kevin Panuelos <kmp091@gmail.com>